Remove .component files
