--- packages/craftcms-cp/src/components/button-group/button-group.css
@@ -1,17 +1,0 @@
-craft-button-group > * {
-  border-radius: 0;
-}
-
-craft-button-group > craft-button:first-child,
-craft-button-group > craft-dropdown:first-child > craft-button,
-craft-button-group > wa-dropdown:first-child > craft-button {
-  border-start-start-radius: var(--c-radius-sm);
-  border-end-start-radius: var(--c-radius-sm);
-}
-
-craft-button-group > craft-button:last-child,
-craft-button-group > wa-dropdown:last-child > craft-button,
-craft-button-group > craft-dropdown:last-child > craft-button {
-  border-start-end-radius: var(--c-radius-sm);
-  border-end-end-radius: var(--c-radius-sm);
-}
