+++ packages/craftcms-cp/src/components/button-group/button-group.ts
@@ -1,7 +1,20 @@
-import CraftButtonGroup from './button-group.component.js';
+import {property} from 'lit/decorators.js';
+import {html, LitElement} from 'lit';
+import styles from './button-group.styles.js';
+import type {CSSResultGroup} from 'lit';
 
-export * from './button-group.component.js';
-export default CraftButtonGroup;
+/**
+ * @summary Wrapper component used to group a set of buttons together.
+ *
+ * @slot - The default slot.
+ */
+export default class CraftButtonGroup extends LitElement {
+  static override styles: CSSResultGroup = [styles];
+
+  override render() {
+    return html`<slot></slot>`;
+  }
+}
 
 if (!customElements.get('craft-button-group')) {
   customElements.define('craft-button-group', CraftButtonGroup);
