+++ packages/craftcms-cp/src/components/input/input.ts
@@ -1,7 +1,25 @@
-import CraftInput from './input.component.js';
+import {LionInput} from '@lion/ui/input.js';
+import {inputStyles} from '../../styles/form.styles.js';
+import styles from './input.styles.js';
+import {property} from 'lit/decorators.js';
 
-export * from './input.component.js';
-export default CraftInput;
+export default class CraftInput extends LionInput {
+  static override get styles() {
+    return [...super.styles, inputStyles, styles];
+  }
+
+  @property({type: Number, reflect: true}) size: string = '';
+
+  override connectedCallback() {
+    super.connectedCallback();
+    if (this._inputNode) {
+      const sizeInt = parseInt(this.size, 10);
+      if (sizeInt > 0) {
+        this._inputNode.size = sizeInt;
+      }
+    }
+  }
+}
 
 if (!customElements.get('craft-input')) {
   customElements.define('craft-input', CraftInput);
