+++ packages/craftcms-cp/src/components/spinner/spinner.ts
@@ -1,8 +1,19 @@
-import CraftSpinner from './spinner.component.js';
+import {LitElement, html, css} from 'lit';
+import {property} from 'lit/decorators.js';
+import componentStyles from './spinner.styles.js';
 
-export * from './spinner.component.js';
+export default class CraftSpinner extends LitElement {
+  static override styles = [componentStyles];
 
-export default CraftSpinner;
+  protected override render() {
+    return html`
+      <div tabindex="-1">
+        <div class="spinner"></div>
+        <span class="message visually-hidden"><slot /></span>
+      </div>
+    `;
+  }
+}
 
 if (!customElements.get('craft-spinner')) {
   customElements.define('craft-spinner', CraftSpinner);
