+++ packages/craftcms-cp/src/components/tab/tab.ts
@@ -1,8 +1,27 @@
-import CraftTab from './tab.component.js';
-import './tab.css';
+import {property} from 'lit/decorators.js';
+import {html, LitElement} from 'lit';
+import styles from './tab.styles.js';
+import type {CSSResultGroup} from 'lit';
 
-export * from './tab.component.js';
-export default CraftTab;
+/**
+ * @summary Short summary of the component's intended use.
+ *
+ * @event craft-event-name - Emitted as an example.
+ *
+ * @slot - The default slot.
+ * @slot example - An example slot.
+ *
+ * @csspart base - The component's base wrapper.
+ *
+ * @cssproperty --example - An example CSS custom property.
+ */
+export default class CraftTab extends LitElement {
+  static override styles: CSSResultGroup = [styles];
+
+  override render() {
+    return html`<slot></slot> `;
+  }
+}
 
 if (!customElements.get('craft-tab')) {
   customElements.define('craft-tab', CraftTab);
