--- packages/craftcms-cp/src/components/tabs/tabs.component.ts
@@ -1,8 +1,0 @@
-import {LionTabs} from '@lion/ui/tabs.js';
-import styles from './tabs.styles.js';
-
-export default class CraftTabs extends LionTabs {
-  static override get styles() {
-    return [...super.styles, styles];
-  }
-}
