+++ packages/craftcms-cp/src/components/tabs/tabs.ts
@@ -1,7 +1,11 @@
-import CraftTabs from './tabs.component.js';
+import {LionTabs} from '@lion/ui/tabs.js';
+import styles from './tabs.styles.js';
 
-export * from './tabs.component.js';
-export default CraftTabs;
+export default class CraftTabs extends LionTabs {
+  static override get styles() {
+    return [...super.styles, styles];
+  }
+}
 
 if (!customElements.get('craft-tabs')) {
   customElements.define('craft-tabs', CraftTabs);
