+++ packages/craftcms-cp/src/components/textarea/textarea.ts
@@ -1,7 +1,12 @@
-import CraftTextarea from './textarea.component.js';
+import styles from './textarea.styles.js';
+import {inputStyles} from '../../styles/form.styles.js';
+import {LionTextarea} from '@lion/ui/textarea.js';
 
-export * from './textarea.component.js';
-export default CraftTextarea;
+export default class CraftTextarea extends LionTextarea {
+  static override get styles() {
+    return [...super.styles, inputStyles, styles];
+  }
+}
 
 if (!customElements.get('craft-textarea')) {
   customElements.define('craft-textarea', CraftTextarea);
